Refactor to simplify interface and navigation
