+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Route, Routes } from "react-router-dom";
-import { Browse } from "./pages/Browse";
+import { EditorPage } from "./pages/EditorPage";
 import { MyDocuments } from "./pages/MyDocuments";
 import styles from './App.module.css';
 import axios from "axios";
@@ -10,7 +10,7 @@ function App() {
   return (
     <>
       <Routes>
-        <Route path="/" element={<Browse />} />
+        <Route path="/editor" element={<EditorPage />} />
         <Route path="/my-documents" element={<MyDocuments />} />
       </Routes>
     </>
